+++ chain/chain/src/spice/core_writer_actor.rs
@@ -1,4 +1,5 @@
-use crate::spice::core::{SpiceCoreReader, all_stake_fallback_assignment};
+use crate::spice::all_stake_fallback::all_stake_fallback_assignment;
+use crate::spice::core::SpiceCoreReader;
 use itertools::Itertools;
 use near_async::messaging::{Handler, Sender};
 use near_cache::SyncLruCache;
