+++ chain/chain/src/spice/mod.rs
@@ -1,3 +1,4 @@
+pub mod all_stake_fallback;
 mod ancestry_endorsements;
 pub mod block_application;
 pub mod chain;
