+++ chain/client/src/spice/data_distributor_actor.rs
@@ -20,7 +20,8 @@ use near_async::messaging::IntoSender;
 use near_async::messaging::Sender;
 use near_async::time::Duration;
 use near_chain::Block;
-use near_chain::spice::core::{SpiceCoreReader, fallback_eligible};
+use near_chain::spice::all_stake_fallback::fallback_eligible;
+use near_chain::spice::core::SpiceCoreReader;
 use near_chain::spice::core_writer_actor::ProcessedBlock;
 use near_chain::stateless_validation::metrics::PROCESS_CONTRACT_CODE_REQUEST_TIME;
 use near_chain_configs::MutableValidatorSigner;
